requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.8.1+cpu
torchvision==0.9.1+cpu
torchaudio==0.8.1
boto3==1.17.84
botocore==1.20.84
certifi==2020.12.5
chardet==4.0.0
idna==2.10
jmespath==0.10.0
numpy==1.20.3
opencv-python==4.5.2.52
pika==1.2.0
Pillow==8.2.0
python-dateutil==2.8.1
python-dotenv==0.17.1
requests==2.25.1
s3transfer==0.4.2
six==1.16.0
typing-extensions==3.10.0.0
urllib3==1.26.4
voluptuous==0.12.1
